// legacy/src/models/enums.ts
export enum UnitOfMeasure {
    Pound = 'lbs',
    Kilogram = 'kgs'
}

export enum ToastType {
    Default = 'default',
    Error = 'error',
    Success = 'success',
    Warning = 'warning'
}

export enum AlertType {
    Default = 'default',
    Error = 'error',
    Success = 'success',
    Warning = 'warning',
    Info = 'info'
}

export enum LoginProvider {
    Google = 'google',
    GitHub = 'github'
}

export enum Entities {
    UserProfile = 'user_profile',
    GymCheckIn = 'gym_checkin'
}